Implemented synchronous PID controller
SynchronousPID provides a Calculate() function for teams to call themselves
instead of running the controller with a Notifier.
