+++ wpilibc/src/main/native/include/PIDInterface.h
@@ -7,11 +7,9 @@
 
 #pragma once
 
-#include "Controller.h"
-
 namespace frc {
 
-class PIDInterface : public Controller {
+class PIDInterface {
   virtual void SetPID(double p, double i, double d) = 0;
   virtual double GetP() const = 0;
   virtual double GetI() const = 0;
@@ -19,10 +17,6 @@ class PIDInterface : public Controller {
 
   virtual void SetSetpoint(double setpoint) = 0;
   virtual double GetSetpoint() const = 0;
-
-  virtual void Enable() = 0;
-  virtual void Disable() = 0;
-  virtual bool IsEnabled() const = 0;
 
   virtual void Reset() = 0;
 };
